@@ -2,6 +2,7 @@
  * Created by linzx on 2017/7/28.
  */
 
+
 /**
  * local     ：本地环境。连接本地的后端接口（127.0.0.1），用于本地测试。         【后端开发人员】、【测试人员】
  * local_test：本地调试。给前端人员本地调试代码用，使用模拟的json数据进行测试    【前端开发人员】
@@ -15,13 +16,13 @@ var api_domain, daily_test, local_test;
 if (active === "product") {
 	api_domain = ""; //线上API端口，待配置
 } else if (active === "daily_ued") {
-	api_domain = "192.168.200.161:8080";
+	 api_domain = "192.168.200.161:8080";
 	// api_domain = "192.168.118.157:8080";//王俊明
 	// api_domain = "192.168.31.149:8080";//鲁艳红
-	// api_/domain = "192.168.31.163:8088"; //轩腾飞
+//	 api_domain = "192.168.31.163:8095"; //轩腾飞
 	// api_domain = "192.168.31.141:8080"; //陈强
 	// api_domain = "192.168.118.93:8083";//李小方
-	// api_domain = "192.168.118.106:8080";//徐俊
+	// api_domain = "192.168.31.143:8080";//徐俊
 
 	// daily_test = true; //日常服务器的标识s
 } else if (active === "daily") {
