@@ -380,17 +380,17 @@ app.controller("configController", ["$rootScope", "$scope", "$http", "$location"
         //设定排程前的顺序 === [颜色信息校验,刀具信息校验,夹具信息校验,产能校验,物料信息校验,PAP信息校验,工艺路线信息校验,排程步骤校验,适宜设备校验,系统配置校验]
         $scope.scheduleFrontData = [];
         var frontData = scheduleTableViewModelService.validation_rules_from(res); //获得排程前需要的数据
-        //"CT_INFO_CHECKING","FIXTURE_INFO_CHECKING","PAP_INFO_CHECKING",
-        var scheduleFrontData = ["COLOR_INFO_CHECKING", "CAPABILITY_INFO_CHECKING", "MATERIAL_INFO_CHECKING", "ROUTING_INFO_CHECKING", "SUITABLE_PRODUCT_INFO_CHECKING", "SYSTEM_CONFIG_CHECKING"];
+        //
+        var scheduleFrontData = ["COLOR_INFO_CHECKING", "CAPABILITY_INFO_CHECKING", "MATERIAL_INFO_CHECKING", "ROUTING_INFO_CHECKING", "SUITABLE_PRODUCT_INFO_CHECKING", "SYSTEM_CONFIG_CHECKING", "CT_INFO_CHECKING", "FIXTURE_INFO_CHECKING", "PAP_INFO_CHECKING"];
         scheduleFrontData.forEach(function (item, index) {
             $scope.scheduleFrontData.push(frontData[item]);
         });
 
-        //设定排程后的顺序 === [物流疏通校验,齐套性校验,组合件校验,超产能校验,生产时间校验]
+        //设定排程后的顺序 === [超产能校验,齐套性校验,组合件校验,生产时间校验]
         $scope.scheduleLaterData = [];
         var laterData = scheduleTableViewModelService.validation_rules_later(res); //获得排程后需要的数据
-        //"PROCESS_SEQ_CHECKING","POMO_SUIT_CHECKING","ASSEMBLING_UNIT",
-        var scheduleLaterData = ["CAPABILITY_OVER_CHECKING", "PRODUCTION_TIME_CHECKING"];
+        //
+        var scheduleLaterData = ["CAPABILITY_OVER_CHECKING", "POMO_SUIT_CHECKING", "ASSEMBLING_UNIT", "PRODUCTION_TIME_CHECKING"];
         scheduleLaterData.forEach(function (item) {
             $scope.scheduleLaterData.push(laterData[item]);
         });
@@ -410,49 +410,6 @@ app.controller("configController", ["$rootScope", "$scope", "$http", "$location"
             });
             //控制选中状态
         }, 0);
-    };
-    //根据此模块设置各个排程规则项的是否要展示
-    $scope.scheduleRuleItemShow = {
-        isLoadOverduePoolTask: true, //当前日期前的车间计划 ：
-        overduePeriod: true, //拉取当前车间计划天数
-        schedulePeriodSelected: true, //排程周期
-        scheduleInterval: true, //排程间隔
-        minScheduleDay: true, //最早起排日期
-        freezePeriod: true, //冻结期
-        //排程前校验
-        preScheduleCheckMap: {
-            CAPABILITY_INFO_CHECKING: false, //产能校验
-            COLOR_INFO_CHECKING: false, //颜色信息校验
-            CT_INFO_CHECKING: true, //刀具信息校验
-            FIXTURE_INFO_CHECKING: true, //夹具信息校验
-            MATERIAL_INFO_CHECKING: false, //物料信息校验
-            PAP_INFO_CHECKING: true, //PAP信息校验
-            ROUTING_INFO_CHECKING: false, //排程步骤校验
-            SUITABLE_PRODUCT_INFO_CHECKING: false, //适宜设备校验
-            SYSTEM_CONFIG_CHECKING: false },
-        //排程后校验
-        postScheduleCheckMap: {
-            ASSEMBLING_UNIT: true, //组合件校验
-            CAPABILITY_OVER_CHECKING: false, //超产能校验
-            POMO_SUIT_CHECKING: true, //齐套性校验
-            PROCESS_SEQ_CHECKING: true, //物流疏通校验
-            PRODUCTION_TIME_CHECKING: false },
-        papTypeSelected: true,
-        schedulePointSelected: true, //起排工序
-        isAheadOn: true, //启用延迟生产
-        isCombinationOn: true, //启用组合件
-        isEconomicOn: true, //启用经济批量
-        isFrequencyOn: true, //启用生产频度
-        isUseNotMassEqu: true, //使用非批量设备
-        //排程因子权重
-        scheduleWeightMap: {
-            KEY_EQUIPMENT: true, //关键设备
-            MOST_USED_EQUIPMENT: true, //高负荷设备
-            SAME_COLOR: true, //相同颜色
-            SAME_CT: true, //相同刀具
-            SAME_FIXTURE: true, //相同物料
-            SAME_MATERIAL: true, //相似物料
-            SIMILAR_MATERIAL: true }
     };
 
     /**
@@ -868,7 +825,6 @@ app.controller("onlineConfigController", ["$rootScope", "$scope", "$state", "too
 
     //获取所有设备
     //设备下拉列表数据
-    $scope.equipmentList = {};
     $scope.equipmentSelectList = [];
     $scope.equipmentData = {
         showText: "设备",
@@ -900,6 +856,11 @@ app.controller("onlineConfigController", ["$rootScope", "$scope", "$state", "too
 
         return $http.get($rootScope.restful_api.search_material_code + params).then(function (res) {
             $scope.materialCodeList = res.data.slice(0, 100);
+            $scope.materialCodeList.unshift({
+                materialCode: "请选择",
+                materialName: "请选择",
+                materialId: ""
+            });
             $scope.materialCodeList.forEach(function (item) {
                 item.text = item.materialCode;
             });
@@ -913,6 +874,11 @@ app.controller("onlineConfigController", ["$rootScope", "$scope", "$state", "too
 
         $http.get($rootScope.restful_api.search_material_name + params).then(function (res) {
             $scope.materialNameList = res.data.slice(0, 100);
+            $scope.materialNameList.unshift({
+                materialCode: "请选择",
+                materialName: "请选择",
+                materialId: ""
+            });
             $scope.materialNameList.forEach(function (item) {
                 item.text = item.materialName;
             });
@@ -927,16 +893,16 @@ app.controller("onlineConfigController", ["$rootScope", "$scope", "$state", "too
 
     //根据改动条件，获得equipment设备列表
     $scope.getEquipment = function () {
-        console.log($scope.equipmentSelectTypeList);
         http.get({
             url: $rootScope.restful_api.get_all_equipment + "?startTime=" + tool.getCorrectDate(new Date()) + "&endTime=" + tool.getCorrectDate(new Date()) + "&searchType=1" + "&modelIdList=" + $scope.equipmentSelectTypeList.join(",") + "&locationFilterList=" + $scope.get_selected_location().join(","),
             successFn: function successFn(res) {
+                $scope.equipmentList = res.data;
                 var productUnit = res.data;
                 var equipmentList = [];
                 for (var i in productUnit) {
                     equipmentList.push({
                         dragItemText: productUnit[i].productUnitName,
-                        id: productUnit[i].productUnitId + "_" + (productUnit[i].type === "PRODUCTION_UNIT" ? 1 : 0)
+                        id: productUnit[i].productUnitId + "_" + productUnit[i].type
                     });
                 }
                 $scope.equipmentData.repeatData = equipmentList;
@@ -948,7 +914,11 @@ app.controller("onlineConfigController", ["$rootScope", "$scope", "$state", "too
         var postData = [];
         //创建好待会需要渲染表格的数组
         $scope.onlineConfigList = [];
-        //坑爹的后台数据，前台加个判断，如果设备不选，默认全部全选
+        /*因为坑爹的后台数据没法一次提供所有需要的，
+         所以要先从选中的设备列表里面，根据设备id_type，到整个设备列表去获得设备名，得出一个新数组$scope.onlineConfigList
+         然后再去到请求获得数据里面根据设备id_type去获取对应的物料信息，将获得到的信息合并到$scope.onlineConfigList
+         */
+        //	前台加个判断，如果设备不选，默认全部全选
         if ($scope.equipmentSelectList.length === 0) {
             for (var i in $scope.equipmentList) {
                 postData.push({
@@ -956,43 +926,30 @@ app.controller("onlineConfigController", ["$rootScope", "$scope", "$state", "too
                     "produceUnitType": $scope.equipmentList[i].type
                 });
             }
-            //坑爹的后台数据，又要从设备列表列表里面再去获得一次设备名
-            postData.forEach(function (selectItem) {
-                $scope.equipmentData.repeatData.some(function (item) {
-                    if (item.id === selectItem.equipmentId + "_" + selectItem.produceUnitType) {
-                        $scope.onlineConfigList.push({
-                            equipmentName: item.dragItemText,
-                            equipmentCode: item.equipmentCode,
-                            equipmentId: selectItem.equipmentId,
-                            equipmentType: selectItem.produceUnitType
-                        });
-                        return true;
-                    }
-                });
-            });
         } else {
             $scope.equipmentSelectList.forEach(function (item) {
                 var arr = item.split("_");
                 postData.push({
                     "equipmentId": arr[0],
-                    "produceUnitType": arr[1]
-                });
-            });
-            //坑爹的后台数据，又要从设备列表列表里面再去获得一次设备名
-            $scope.equipmentSelectList.forEach(function (selectItem) {
-                $scope.equipmentData.repeatData.some(function (item) {
-                    if (item.id === selectItem) {
-                        $scope.onlineConfigList.push({
-                            equipmentName: item.dragItemText,
-                            equipmentCode: item.equipmentCode,
-                            equipmentId: selectItem.equipmentId,
-                            equipmentType: selectItem.produceUnitType
-                        });
-                        return true;
-                    }
+                    "produceUnitType": item.replace(arr[0] + "_", "")
                 });
             });
         }
+        //从选中的设备列表里面，根据设备id_type，到整个设备列表去获得设备名，得出一个新数组$scope.onlineConfigList
+        postData.forEach(function (selectItem) {
+            $scope.equipmentData.repeatData.some(function (item) {
+                if (item.id === selectItem.equipmentId + "_" + selectItem.produceUnitType) {
+                    $scope.onlineConfigList.push({
+                        equipmentName: item.dragItemText,
+                        equipmentCode: item.equipmentCode,
+                        equipmentId: selectItem.equipmentId,
+                        equipmentType: selectItem.produceUnitType
+                    });
+                    return true;
+                }
+            });
+        });
+        // 请求获得数据里面根据设备id_type去获取对应的物料信息，将获得到的信息合并到$scope.onlineConfigList
         http.post({
             url: $rootScope.restful_api.search_online_config,
             data: postData,
@@ -1001,23 +958,18 @@ app.controller("onlineConfigController", ["$rootScope", "$scope", "$state", "too
                     //获得渲染表格所需要的数据
                     $scope.onlineConfigList.forEach(function (onlineItem) {
                         if (resItem.equipmentId + "_" + resItem.produceUnitType === onlineItem.equipmentId + "_" + onlineItem.equipmentType) {
-                            // ({
-                            // 	materialCode:onlineItem.materialCode,
-                            // 	materialId:onlineItem.materialId,
-                            // 	materialName:onlineItem.materialName
-                            // } = resItem);
                             onlineItem.materialCode = resItem.materialCode;
                             onlineItem.materialId = resItem.materialId;
                             onlineItem.materialName = resItem.materialName;
                         }
-                        // //获得生产物料模糊搜索需要的下拉列表数据
-                        // onlineItem.materialCodeList = $scope.materialCodeList;
-                        // onlineItem.materialNameList = $scope.materialNameList;
                     });
                 });
-
                 //出现table表格和按钮
                 $scope.isShowSearchConfig = true;
+            },
+            errorFn: function errorFn(res) {
+                //table表格和按钮消失
+                $scope.isShowSearchConfig = false;
             }
         });
     };
@@ -1062,13 +1014,13 @@ app.controller("onlineConfigController", ["$rootScope", "$scope", "$state", "too
     $scope.saveOnlineConfig = function () {
         var postData = [];
         $scope.onlineConfigList.forEach(function (item) {
-            if (item.materialId) {
-                postData.push({
-                    equipmentId: Number(item.equipmentId),
-                    produceUnitType: item.equipmentType,
-                    materialId: item.materialId
-                });
-            }
+            // if(item.materialId){
+            postData.push({
+                equipmentId: Number(item.equipmentId),
+                produceUnitType: item.equipmentType,
+                materialId: item.materialId
+            });
+            // }
         });
         http.post({
             url: $rootScope.restful_api.save_online_config,
@@ -1326,7 +1278,7 @@ app.controller("planController", ["$rootScope", "$scope", "$timeout", "$q", "sch
      * desc:根据点击车间，改变中间车间列表展示数据，需要展示哪些车间
      * time:2017-06-12
      * last:linzx
-     * @param: locationId :车间id，
+     * @param: locationId :点击的车间id，
      **/
     var changeParentLocationList = function changeParentLocationList(locationId) {
         var parentLocationId = locationId.slice(0, -2);
@@ -1375,87 +1327,6 @@ app.controller("planController", ["$rootScope", "$scope", "$timeout", "$q", "sch
         //如果父元素车间不是最上级车间，则继续向上改变
         if (parentLocationId && parentLocationId !== "01") {
             changeParentLocationList(parentLocationId);
-        }
-    };
-
-    /**排程方案保存进行发送数据**/
-    $scope.saveSchedulePlan = function () {
-        var planNameVal = $(".jPlanName input").val(),
-            currentPlanLi = $(".schedulePlanLi.active"),
-            locationLi = $("#workshopRule").children("ul").children("li");
-        if (!planNameVal.length) {
-            layer.alert("排程方案名不可以为空");
-            return;
-        }
-        //如果保存方案中时没有车间，则给出提示
-        if (locationLi.length <= 0) {
-            layer.alert("请为排程方案至少选择一个车间");
-            return;
-        }
-        //新建一个需要发送的post数据
-        var postData = {
-            "schemeId": currentPlanLi.attr("data-scheme-id"),
-            "schemeName": planNameVal,
-            "locationRuleList": []
-        };
-        var emptyRule = true; //用于后面判断
-        //遍历所有的车间li用于获取车间顺序和相应的规则信息
-        locationLi.each(function () {
-            var item = $(this)[0];
-            var locationA = item.firstElementChild;
-            var ruleA = item.lastElementChild.firstElementChild;
-            var obj = {
-                "locationId": item.getAttribute("data-location-id"),
-                "locationName": locationA.innerHTML,
-                "ruleId": ruleA.getAttribute("data-rule-id"),
-                "ruleName": ruleA.innerHTML
-            };
-            if (ruleA.getAttribute("data-rule-id") === "") {
-                emptyRule = false;
-            }
-            postData.locationRuleList.push(obj);
-        });
-        if (!emptyRule) {
-            layer.alert("请为每个排程车间配备排程规则");
-            return;
-        }
-        /*判断是否是临时方案，如果是更新，不是则新建方案
-         *true : 临时方案，执行新建操作；false ：已保存方案，执行更新
-        */
-        if (postData.schemeId.slice(0, 9) === "temporary") {
-            delete postData.schemeId;
-            http.post({
-                url: $rootScope.restful_api.single_schedule_plan,
-                data: postData,
-                successFn: function successFn(res) {
-                    if (res.data) {
-                        $scope.info.success("排程方案新建成功");
-                        //将临时id替换成获得的方案id
-                        currentPlanLi.attr("data-scheme-id", res.data);
-                        $scope.currentPlan.schemeId = res.data;
-                        $scope.schemeId = res.data;
-                    }
-                    /*
-                     * 判断是否临时创建的方案，用于离开排程方案页面进行判断，
-                     * 判断代码写在路由配置文件里routerConfig.js
-                     */
-                    sessionStorage.temporarySchemeLength = Object.keys($scope.temporarySchemeData).length;
-                },
-                errorFn: function errorFn() {
-                    $scope.info.fail("排程方案新建失败，请检查服务器");
-                }
-            });
-        } else {
-            http.put({
-                url: $rootScope.restful_api.single_schedule_plan + $scope.schemeId,
-                data: postData,
-                successFn: function successFn(res) {
-                    $scope.info.success("排程方案更新成功");
-                },
-                errorFn: function errorFn() {
-                    $scope.info.fail("排程方案更新失败");
-                }
-            });
         }
     };
 
@@ -1532,6 +1403,87 @@ app.controller("planController", ["$rootScope", "$scope", "$timeout", "$q", "sch
         */
         sessionStorage.temporarySchemeLength = Object.keys($scope.temporarySchemeData).length;
         layer.closeAll();
+    };
+
+    /**排程方案保存进行发送数据**/
+    $scope.saveSchedulePlan = function () {
+        var planNameVal = $(".jPlanName input").val(),
+            currentPlanLi = $(".schedulePlanLi.active"),
+            locationLi = $("#workshopRule").children("ul").children("li");
+        if (!planNameVal.length) {
+            layer.alert("排程方案名不可以为空");
+            return;
+        }
+        //如果保存方案中时没有车间，则给出提示
+        if (locationLi.length <= 0) {
+            layer.alert("请为排程方案至少选择一个车间");
+            return;
+        }
+        //新建一个需要发送的post数据
+        var postData = {
+            "schemeId": currentPlanLi.attr("data-scheme-id"),
+            "schemeName": planNameVal,
+            "locationRuleList": []
+        };
+        var emptyRule = true; //用于后面判断
+        //遍历所有的车间li用于获取车间顺序和相应的规则信息
+        locationLi.each(function () {
+            var item = $(this)[0];
+            var locationA = item.firstElementChild;
+            var ruleA = item.lastElementChild.firstElementChild;
+            var obj = {
+                "locationId": item.getAttribute("data-location-id"),
+                "locationName": locationA.innerHTML,
+                "ruleId": ruleA.getAttribute("data-rule-id"),
+                "ruleName": ruleA.innerHTML
+            };
+            if (ruleA.getAttribute("data-rule-id") === "") {
+                emptyRule = false;
+            }
+            postData.locationRuleList.push(obj);
+        });
+        if (!emptyRule) {
+            layer.alert("请为每个排程车间配备排程规则");
+            return;
+        }
+        /*判断是否是临时方案，如果是更新，不是则新建方案
+         *true : 临时方案，执行新建操作；false ：已保存方案，执行更新
+         */
+        if (postData.schemeId.slice(0, 9) === "temporary") {
+            delete postData.schemeId;
+            http.post({
+                url: $rootScope.restful_api.single_schedule_plan,
+                data: postData,
+                successFn: function successFn(res) {
+                    if (res.data) {
+                        $scope.info.success("排程方案新建成功");
+                        //将临时id替换成获得的方案id
+                        currentPlanLi.attr("data-scheme-id", res.data);
+                        $scope.currentPlan.schemeId = res.data;
+                        $scope.schemeId = res.data;
+                    }
+                    /*
+                     * 判断是否临时创建的方案，用于离开排程方案页面进行判断，
+                     * 判断代码写在路由配置文件里routerConfig.js
+                     */
+                    sessionStorage.temporarySchemeLength = Object.keys($scope.temporarySchemeData).length;
+                },
+                errorFn: function errorFn() {
+                    $scope.info.fail("排程方案新建失败，请检查服务器");
+                }
+            });
+        } else {
+            http.put({
+                url: $rootScope.restful_api.single_schedule_plan + $scope.schemeId,
+                data: postData,
+                successFn: function successFn(res) {
+                    $scope.info.success("排程方案更新成功");
+                },
+                errorFn: function errorFn() {
+                    $scope.info.fail("排程方案更新失败");
+                }
+            });
+        }
     };
 
     //查看方案
@@ -1776,7 +1728,7 @@ app.controller("ruleController", ["$rootScope", "$scope", "$http", "$timeout", "
                 http.get({
                     url: $rootScope.restful_api.single_schedule_rule + $scope.ruleId,
                     successFn: function successFn(res) {
-                        $scope.rule = $scope.ruleList[0];
+                        $scope.currentRule = $scope.ruleList[0];
                         $scope.renderRulePage(res.data);
                     }
                 })
@@ -1905,6 +1857,64 @@ app.controller("ruleController", ["$rootScope", "$scope", "$http", "$timeout", "
         return $scope.ruleNav.active === val;
     };
 
+    //弹出新建排程规则窗口
+    $scope.openRuleWindow = function () {
+        var index = layer.open({
+            type: 1,
+            title: false,
+            closeBtn: 0,
+            shadeClose: true,
+            skin: 'yourclass',
+            content: $(".addNewRule"),
+            success: function success() {
+                $(".addNewRule").on("click", ".close,.cancal", function () {
+                    layer.close(index);
+                });
+            }
+        });
+    };
+
+    //创建临时规则数据库，存储临时创建但是未保存的规则，刷新或者离开页面数据库消失，表示用户不想保存
+    $scope.temporaryRuleData = {};
+    //执行临时创建规则
+    $scope.recordTemporaryPlan = function () {
+        //检测是否重名
+        if (!$scope.newRuleName) {
+            layer.msg('请输入正确的排程规则名');
+            return false;
+        }
+        if (!tool.checkRepeatName($scope.newRuleName, $scope.ruleList, "rule")) {
+            layer.msg('排程规则名重复，请重新输入');
+            return;
+        }
+        layer.closeAll();
+        //判断用户是否选了规则,没选设置为默认规则
+        $scope.ruleId = $("#ruleSelect").val() === "0" ? "default" : $("#ruleSelect").val();
+        http.get({
+            url: $rootScope.restful_api.single_schedule_rule + $scope.ruleId,
+            successFn: function successFn(res) {
+                $scope.scheduleCheckData = res.data;
+                $scope.ruleId = "temporary" + new Date().getTime();
+                $scope.currentRule = {
+                    "ruleId": $scope.ruleId,
+                    "ruleName": $scope.newRuleName
+                };
+                $scope.ruleList.push($scope.currentRule);
+                $scope.temporaryRuleData[$scope.ruleId] = $scope.scheduleCheckData;
+
+                $scope.renderRulePage($scope.scheduleCheckData);
+
+                //最后一个li获得actice的状态
+                $timeout(function () {
+                    $(".check-rule-nav .rule-li").last().addClass("active").siblings().removeClass("active");
+                }, 0);
+            },
+            errorFn: function errorFn() {
+                $scope.info.fail("获取排程规则失败，请检查网络");
+            }
+        });
+    };
+
     /**点击保存排程规则**/
     $scope.saveCheckConfig = function () {
         //排程规则名字不可为空
@@ -1948,105 +1958,60 @@ app.controller("ruleController", ["$rootScope", "$scope", "$http", "$timeout", "
         post.minScheduleDay = $("#minScheduleDay").val();
         //获取规则名字
         post.ruleName = ruleName;
-        //发送数据
-        http.put({
-            url: $rootScope.restful_api.single_schedule_rule + $scope.ruleId,
-            data: post,
-            successFn: function successFn(res) {
-                if (res.data.error_response) {
-                    $scope.info.fail("数据保存失败");
-                } else {
-                    $scope.info.success("数据保存成功");
-                }
-            },
-            errorFn: function errorFn() {
-                $scope.info.fail("数据保存失败,请检查服务器");
-            }
-        });
-    };
 
-    //弹出新建排程规则窗口
-    $scope.openRuleWindow = function () {
-        var index = layer.open({
-            type: 1,
-            title: false,
-            closeBtn: 0,
-            shadeClose: true,
-            skin: 'yourclass',
-            content: $(".addNewRule"),
-            success: function success() {
-                $(".addNewRule").on("click", ".close,.cancal", function () {
-                    layer.close(index);
-                });
-            }
-        });
-    };
-
-    //发送新建规则请求
-    $scope.postNewRule = function () {
-        //检测是否重名
-        if (!$scope.newRuleName) {
-            layer.msg('请输入正确的排程规则名');
-            return false;
-        }
-        if (!tool.checkRepeatName($scope.newRuleName, $scope.ruleList, "rule")) {
-            layer.msg('排程规则名重复，请重新输入');
-            return;
-        }
-        //判断用户是否选了规则,没选设置为默认规则
-        $scope.ruleId = $("#ruleSelect").val() === "0" ? "default" : $("#ruleSelect").val();
-        http.get({
-            url: $rootScope.restful_api.single_schedule_rule + $scope.ruleId,
-            successFn: function successFn(res) {
-                $scope.scheduleCheckData = res.data;
-                $scope.info.success("创建排程规则成功");
-            },
-            errorFn: function errorFn() {
-                $scope.info.fail("创建排程规则失败，请检查网络");
-            }
-        })
-        //再获得对应的数据,then create Rule
-        .then(function () {
-            //删除获取到的Id,post创建规则
-            delete $scope.scheduleCheckData.ruleId;
-            $scope.scheduleCheckData.ruleName = $scope.newRuleName;
+        //判断是新建规则还是更新规则
+        if (($scope.ruleId + "").slice(0, 9) === "temporary") {
+            delete post.ruleId;
+            //新建
             http.post({
                 url: $rootScope.restful_api.single_schedule_rule,
-                data: $scope.scheduleCheckData,
+                data: post,
                 successFn: function successFn(res) {
-                    res = res.data;
-                    $scope.ruleId = res;
-                    $scope.ruleList.push({
-                        "ruleId": res,
-                        "ruleName": $scope.scheduleCheckData.ruleName
-                    });
-                    $scope.renderRulePage($scope.scheduleCheckData);
-                    $scope.newRuleName = "";
-                    layer.closeAll();
-                    //最后一个li获得actice的状态
-                    $timeout(function () {
-                        $scope.lookRule($scope.ruleList[$scope.ruleList.length - 1]);
-                        $(".check-rule-nav .rule-li").last().addClass("active").siblings().removeClass("active");
-                    }, 0);
+                    if (res.data.error_response) {
+                        $scope.info.fail("创建排程规则失败，请检查服务器");
+                    } else {
+                        $scope.ruleId = $scope.currentRule.ruleId = res.data;
+                        $scope.info.success("创建排程规则成功");
+                    }
                 },
                 errorFn: function errorFn() {
-                    $scope.info.fail("创建规则失败-post");
+                    $scope.info.fail("创建排程规则失败，请检查服务器");
                 }
             });
-        });
+        } else {
+            //更新
+            http.put({
+                url: $rootScope.restful_api.single_schedule_rule + $scope.ruleId,
+                data: post,
+                successFn: function successFn(res) {
+                    if (res.data.error_response) {
+                        $scope.info.fail("排程规则更新失败");
+                    } else {
+                        $scope.info.success("排程规则更新成功");
+                    }
+                },
+                errorFn: function errorFn() {
+                    $scope.info.fail("排程规则更新失败,请检查服务器");
+                }
+            });
+        }
     };
 
     //查看规则
     $scope.lookRule = function (rule) {
-        $scope.rule = rule;
-        http.get({
-            url: $rootScope.restful_api.single_schedule_rule + rule.ruleId,
-            successFn: function successFn(res) {
-                $scope.ruleId = rule.ruleId;
-                //渲染规则页面数据
-                $scope.renderRulePage(res.data);
-            }
-        });
+        $scope.currentRule = rule;
+        $scope.ruleId = rule.ruleId;
+        if ((rule.ruleId + "").slice(0, 9) === "temporary") {
+            $scope.renderRulePage($scope.temporaryRuleData[rule.ruleId]);
+        } else {
+            http.get({
+                url: $rootScope.restful_api.single_schedule_rule + rule.ruleId,
+                successFn: function successFn(res) {
+                    //渲染规则页面数据
+                    $scope.renderRulePage(res.data);
+                }
+            });
+        }
     };
 
     //删除排程规则
@@ -2055,47 +2020,61 @@ app.controller("ruleController", ["$rootScope", "$scope", "$http", "$timeout", "
         var deleteRule = layer.confirm('确定删除此排程规则？', {
             btn: ['确定', '取消'] //按钮
         }, function () {
-            var target = event.target || event.srcElement;
-            var parentElement = target.parentNode;
-            var ruleId = parentElement.getAttribute("data-rule-id");
+            var ruleId = $(event.target).parent().attr("data-rule-id");
             layer.close(deleteRule);
-            http.delete({
-                url: $rootScope.restful_api.single_schedule_rule + ruleId,
-                successFn: function successFn(res) {
-                    //成功返回
-                    if (res.data == "1") {
-                        $scope.ruleList.forEach(function (item, index) {
-                            if (item.ruleId == ruleId) {
-                                $scope.ruleList.splice(index, 1);
+            if ((ruleId + "").slice(0, 9) === "temporary") {
+                $timeout(function () {
+
+                    delete $scope.temporaryRuleData[ruleId];
+                    $scope.ruleList = $scope.ruleList.filter(function (item) {
+                        return item.ruleId !== ruleId;
+                    });
+
+                    //没有方案时，不需要执行
+                    if ($scope.ruleList.length !== 0) {
+                        $scope.lookRule($scope.ruleList[0]);
+                    }
+                });
+            } else {
+                http.delete({
+                    url: $rootScope.restful_api.single_schedule_rule + ruleId,
+                    successFn: function successFn(res) {
+                        //成功返回
+                        if (res.data == "1") {
+                            $scope.ruleList.forEach(function (item, index) {
+                                if (item.ruleId == ruleId) {
+                                    $scope.ruleList.splice(index, 1);
+                                }
+                            });
+                            $scope.info.success("删除排程规则成功");
+                            //删除成功，查看第一个排程规则，没有规则时，不需要执行
+                            if ($scope.ruleList.length != 0) {
+                                $timeout(function () {
+                                    $(".check-rule-nav .rule-li").first().addClass("active").siblings().removeClass("active");
+                                }, 0);
+                                $scope.lookRule($scope.ruleList[0]);
                             }
-                        });
-                        $scope.info.success("删除排程规则成功");
-                        //删除成功，查看第一个排程规则，没有规则时，不需要执行
-                        if ($scope.ruleList.length != 0) {
-                            $timeout(function () {
-                                $(".check-rule-nav .rule-li").first().addClass("active").siblings().removeClass("active");
-                            }, 0);
-                            $scope.lookRule($scope.ruleList[0]);
+                        }
+                    },
+                    errorFn: function errorFn(res) {
+                        if (res.data.error_response.code === 102) {
+                            $scope.info.fail("你没有权限删除此规则");
+                            // return
+                        } else if (res.data.error_response.code === 103) {
+                            $scope.info.fail("此规则正在使用，无法删除");
+                            // return
+                        } else {
+                            $scope.info.fail("删除规则失败，请检查服务器");
                         }
                     }
-                },
-                errorFn: function errorFn(res) {
-                    if (res.data.error_response.code === 102) {
-                        $scope.info.fail("你没有权限删除此规则");
-                        // return
-                    } else if (res.data.error_response.code === 103) {
-                        $scope.info.fail("此规则正在使用，无法删除");
-                        // return
-                    } else {
-                        $scope.info.fail("删除规则失败，请检查服务器");
-                    }
-                }
-            });
+                });
+            }
         }, function () {
             layer.close(deleteRule);
         });
     };
 
+    //==========================上线配置相关代码---start-=================//
     //查看设备初始状态
     //获取所有的地点
     http.get({
@@ -2191,8 +2170,9 @@ app.controller("ruleController", ["$rootScope", "$scope", "$http", "$timeout", "
         $scope.onlineConfigList = [];
         /*因为坑爹的后台数据没法一次提供所有需要的，
          所以要先从选中的设备列表里面，根据设备id_type，到整个设备列表去获得设备名，得出一个新数组$scope.onlineConfigList
-         然后再去到请求获得数据里面根据设备id_type去获取对应的物料信息，
+         然后再去到请求获得数据里面根据设备id_type去获取对应的物料信息，将获得到的信息合并到$scope.onlineConfigList
          */
+        //	前台加个判断，如果设备不选，默认全部全选
         if ($scope.equipmentSelectList.length === 0) {
             for (var i in $scope.equipmentList) {
                 postData.push({
@@ -2200,44 +2180,30 @@ app.controller("ruleController", ["$rootScope", "$scope", "$http", "$timeout", "
                     "produceUnitType": $scope.equipmentList[i].type
                 });
             }
-            //坑爹的后台数据，又要从设备列表列表里面再去获得一次设备名
-            postData.forEach(function (selectItem) {
-                $scope.equipmentData.repeatData.some(function (item) {
-                    if (item.id === selectItem.equipmentId + "_" + selectItem.produceUnitType) {
-                        $scope.onlineConfigList.push({
-                            equipmentName: item.dragItemText,
-                            equipmentCode: item.equipmentCode,
-                            equipmentId: item.id.split("_")[0],
-                            equipmentType: item.id.split("_")[1]
-                        });
-                        return true;
-                    }
-                });
-            });
         } else {
             $scope.equipmentSelectList.forEach(function (item) {
                 var arr = item.split("_");
                 postData.push({
                     "equipmentId": arr[0],
-                    "pUnitType": arr[1]
-                });
-            });
-            //坑爹的后台数据，又要从设备列表列表里面再去获得一次设备名
-            $scope.equipmentSelectList.forEach(function (selectItem) {
-                $scope.equipmentData.repeatData.some(function (item) {
-                    if (item.id === selectItem) {
-                        $scope.onlineConfigList.push({
-                            equipmentName: item.dragItemText,
-                            equipmentCode: item.equipmentCode,
-                            equipmentId: item.id.split("_")[0],
-                            equipmentType: item.id.split("_")[1]
-                        });
-                        return true;
-                    }
+                    "produceUnitType": item.replace(arr[0] + "_", "")
                 });
             });
         }
-        //第一步筛选出用户选中的数据
+        //从选中的设备列表里面，根据设备id_type，到整个设备列表去获得设备名，得出一个新数组$scope.onlineConfigList
+        postData.forEach(function (selectItem) {
+            $scope.equipmentData.repeatData.some(function (item) {
+                if (item.id === selectItem.equipmentId + "_" + selectItem.produceUnitType) {
+                    $scope.onlineConfigList.push({
+                        equipmentName: item.dragItemText,
+                        equipmentCode: item.equipmentCode,
+                        equipmentId: selectItem.equipmentId,
+                        equipmentType: selectItem.produceUnitType
+                    });
+                    return true;
+                }
+            });
+        });
+        // 请求获得数据里面根据设备id_type去获取对应的物料信息，将获得到的信息合并到$scope.onlineConfigList
         http.post({
             url: $rootScope.restful_api.search_online_config,
             data: postData,
@@ -2246,11 +2212,6 @@ app.controller("ruleController", ["$rootScope", "$scope", "$http", "$timeout", "
                     //获得渲染表格所需要的数据
                     $scope.onlineConfigList.forEach(function (onlineItem) {
                         if (resItem.equipmentId + "_" + resItem.produceUnitType === onlineItem.equipmentId + "_" + onlineItem.equipmentType) {
-                            // ({
-                            // 	materialCode:onlineItem.materialCode,
-                            // 	materialId:onlineItem.materialId,
-                            // 	materialName:onlineItem.materialName
-                            // } = resItem);
                             onlineItem.materialCode = resItem.materialCode;
                             onlineItem.materialId = resItem.materialId;
                             onlineItem.materialName = resItem.materialName;
@@ -2262,6 +2223,7 @@ app.controller("ruleController", ["$rootScope", "$scope", "$http", "$timeout", "
             }
         });
     };
+    //==========================上线配置相关代码---end-=================//
 
     //显示规则时，js设置目录栏高度,使左边目录栏的高度可以撑满整个div
     //展开规则时，实时获取规则高度设置，收起时，保证不小于规则最小区域
@@ -2290,60 +2252,69 @@ app.controller("ruleController", ["$rootScope", "$scope", "$http", "$timeout", "
     };
     //显示隐藏高级规则---end
 
-    // //取消绑定，执行unbindWatchPap()
-    // let unbindWatchPap = $scope.$watch("scheduleFrontData[5].select",function (newValue,oldValue) {
-    //     //设置为true，启用默认规则
-    //     if(newValue === true){
-    //         $scope.disable.schedulePointSelected = false;
-    //         $scope.notEdit.schedulePointSelected = true;
-    //         $(".pap-type .scheduleDrag li").eq(2).trigger("click");
-    //     }else {
-    //         //设置为false，下拉选项设置为不启用
-    //         $scope.disable.schedulePointSelected = true;
-    //         $scope.notEdit.schedulePointSelected = true;
-    //         $(".pap-type .scheduleDrag li").eq(0).trigger("click");
-    //     }
-    // });
-    // $("body")
-    // //点击pap类型决定起排工序,和影响排程前校验中的pap检验项
-    //     .on("click",".pap-type li",function(){
-    //         if($(this).attr("data-value")=="PAP_DISABLE"){
-    //             //不启用
-    //             $timeout(function(){
-    //                 $scope.disable.schedulePointSelected = true;
-    //                 $scope.notEdit.schedulePointSelected = true;
-    //                 $scope.scheduleFrontData[5].select = false;//联动排程前校验的pap
-    //                 $(".schedule-point li").eq(0).trigger("click");
-    //             })
-    //         }else if($(this).attr("data-value")=="PAP_SCHEDULE_RULE"){
-    //             $timeout(function(){
-    //                 $scope.disable.schedulePointSelected = false;
-    //                 $scope.notEdit.schedulePointSelected = true;
-    //                 $scope.scheduleFrontData[5].select = true;//联动排程前校验的pap
-    //                 $(".schedule-point li").eq(2).trigger("click");
-    //             })
-    //         }else{
-    //             $timeout(function(){
-    //                 $scope.disable.schedulePointSelected = false;
-    //                 $scope.notEdit.schedulePointSelected = false;
-    //                 $scope.scheduleFrontData[5].select = true;//联动排程前校验的pap
-    //             })
-    //         }
-    //     })
-    //     //排程周期决定排程间隔
-    //     .on("click",".schedulePeriod li",function(){
-    //         if($(this).attr("data-value")=="BY_DAY"){
-    //             $timeout(function(){
-    //                 $scope.disable.scheduleInterval = false;
-    //                 $scope.notEdit.scheduleInterval = false;
-    //             })
-    //         }else{
-    //             $timeout(function(){
-    //                 $scope.disable.scheduleInterval = true;
-    //                 $scope.notEdit.scheduleInterval = true;
-    //             })
-    //         }
-    //     })
+    //取消绑定，执行unbindWatchPap()
+    var unbindWatchPap = $scope.$watch("scheduleFrontData[8].select", function (newValue, oldValue) {
+        //设置为true，启用默认规则
+        if (newValue === true) {
+            $scope.disable.schedulePointSelected = false;
+            $scope.notEdit.schedulePointSelected = true;
+            $(".pap-type .scheduleDrag li").eq(2).trigger("click");
+        } else {
+            //设置为false，下拉选项设置为不启用
+            $scope.disable.schedulePointSelected = true;
+            $scope.notEdit.schedulePointSelected = true;
+            $(".pap-type .scheduleDrag li").eq(0).trigger("click");
+        }
+    });
+    $("body")
+    //点击pap类型决定起排工序,和影响排程前校验中的pap检验项
+    .on("click", ".pap-type li", function () {
+        //获得pap在数组中的位置下标
+        var papIndex = void 0;
+        $scope.scheduleFrontData.some(function (item, i) {
+            if (item.name === 'PAP_INFO_CHECKING') {
+                papIndex = i;
+                return true;
+            }
+            return false;
+        });
+        if ($(this).attr("data-value") == "PAP_DISABLE") {
+            //不启用
+            $timeout(function () {
+                $scope.disable.schedulePointSelected = true;
+                $scope.notEdit.schedulePointSelected = true;
+                $scope.scheduleFrontData[papIndex].select = false; //联动排程前校验的pap
+                $(".schedule-point li").eq(0).trigger("click");
+            });
+        } else if ($(this).attr("data-value") == "PAP_SCHEDULE_RULE") {
+            $timeout(function () {
+                $scope.disable.schedulePointSelected = false;
+                $scope.notEdit.schedulePointSelected = true;
+                $scope.scheduleFrontData[papIndex].select = true; //联动排程前校验的pap
+                $(".schedule-point li").eq(2).trigger("click");
+            });
+        } else {
+            $timeout(function () {
+                $scope.disable.schedulePointSelected = false;
+                $scope.notEdit.schedulePointSelected = false;
+                $scope.scheduleFrontData[papIndex].select = true; //联动排程前校验的pap
+            });
+        }
+    })
+    //排程周期决定排程间隔
+    .on("click", ".schedulePeriod li", function () {
+        if ($(this).attr("data-value") == "BY_DAY") {
+            $timeout(function () {
+                $scope.disable.scheduleInterval = false;
+                $scope.notEdit.scheduleInterval = false;
+            });
+        } else {
+            $timeout(function () {
+                $scope.disable.scheduleInterval = true;
+                $scope.notEdit.scheduleInterval = true;
+            });
+        }
+    });
 }]);
 
 /**
